Apply formatting to value and percent value columns using gt::fmt_number() and gt::fmt_percent()

fmt_acs_values() is a wrapper for gt::fmt_number(), gt::fmt_percent(), and gt::cols_merge_uncert(). If value_col starts with the string supplied to perc_prefix, the value column is formatted as a percent value.

Usage

fmt_acs_values(
  data,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  decimals = 0,
  use_seps = TRUE,
  accounting = FALSE,
  scale_by = 1,
  merge_moe = TRUE,
  currency_value = FALSE,
  ...,
  .cols_fn = starts_with
)

Arguments

data

The gt table data object

obj:<gt_tbl> // required

This is the gt table object that is commonly created through use of the gt() function.

perc_prefix

Prefix string for percent value columns.

perc_sep

Separator string between perc_prefix and the value_col and moe_col strings.

decimals

Number of decimal places

scalar<numeric|integer>(val>=0) // default: 2

This corresponds to the exact number of decimal places to use. A value such as 2.34 can, for example, be formatted with 0 decimal places and it would result in "2". With 4 decimal places, the formatted value becomes "2.3400". The trailing zeros can be removed with drop_trailing_zeros = TRUE. If you always need decimals = 0, the fmt_integer() function should be considered.

use_seps

Use digit group separators

scalar<logical> // default: TRUE

An option to use digit group separators. The type of digit group separator is set by sep_mark and overridden if a locale ID is provided to locale. This setting is TRUE by default.

accounting

Use accounting style

scalar<logical> // default: FALSE

An option to use accounting style for values. Normally, negative values will be shown with a minus sign but using accounting style will instead put any negative values in parentheses.

scale_by

Scale values by a fixed multiplier

scalar<numeric|integer> // default: 1

All numeric values will be multiplied by the scale_by value before undergoing formatting. Since the default value is 1, no values will be changed unless a different multiplier value is supplied. This value will be ignored if using any of the suffixing options (i.e., where suffixing is not set to FALSE).

merge_moe

If TRUE, use gt::cols_merge_uncert() to merge the value_col and moe_col and the percent value and margin of error columns.

currency_value

If TRUE, use gt::fmt_currency() to format value columns instead of gt::fmt_number().

...

Additional parameters passed to gt::fmt_currency() if currency_value = TRUE

.cols_fn

tidyselect function used to select columns to format and merge. Defaults to tidyselect::starts_with