Format American Community Survey estimate and percent estimate columns for a gt table

Create or format a gt table with an estimate and margin of error and (optionally) percent estimate and margin of error value. Use in combination with the select_acs_cols() helper function to prep data before creating a table.

Usage

gt_acs(
  data,
  est_cols = c("estimate", "moe"),
  est_col_label = "Est.",
  perc_cols = c("perc_estimate", "perc_moe"),
  perc_col_label = "% share",
  est_spanner = NULL,
  perc_spanner = NULL,
  combined_spanner = NULL,
  source_note = NULL,
  survey = "acs5",
  year = 2021,
  prefix = "Source: ",
  before = "",
  after = ".",
  table = NULL,
  decimals = 0,
  column_title_col = "column_title",
  column_title_label = NULL,
  append_note = FALSE,
  ...
)

Arguments

data

Input data table

obj:<data.frame>|obj:<tbl_df> // required

A data.frame object or a tibble (tbl_df).

est_cols, est_col_label, est_spanner

Passed to columns, col_labels, and spanner parameter for fmt_acs_estimate()

perc_cols, perc_col_label, perc_spanner

Passed to columns, col_labels, and spanner parameter for fmt_acs_percent()

combined_spanner

If not NULL, combined_spanner is passed to label parameter of gt::tab_spanner() using both est_cols and perc_cols as the columns parameter.

source_note

Source note text

scalar<character> // required

Text to be used in the source note. We can optionally use the md() and html() functions to style the text as Markdown or to retain HTML elements in the text.

survey

ACS survey, "acs5", "acs3", or "acs1".

year

Based on the year and survey, acs_survey_ts() returns a vector of years for non-overlapping ACS samples to allow comparison.

prefix

Text to insert before ACS survey label.

before, after

A character string to be added before/after each word.

table

One or more table IDs to include in label or source note.

decimals

Number of decimal places

scalar<numeric|integer>(val>=0) // default: 2

This corresponds to the exact number of decimal places to use. A value such as 2.34 can, for example, be formatted with 0 decimal places and it would result in "2". With 4 decimal places, the formatted value becomes "2.3400". The trailing zeros can be removed with drop_trailing_zeros = TRUE. If you always need decimals = 0, the fmt_integer() function should be considered.

column_title_col

Column name to check for in the input gt table. Only required if column_title_label is supplied.

column_title_label

Optional label applied to column named in column_title_col.

append_note

If TRUE, add source_note to the end of the generated ACS data label. If FALSE, any supplied source_note will be used instead of an ACS label.

...

Additional parameters passed to gt::fmt_number() by fmt_acs_estimate() or to gt::fmt_percent() by fmt_acs_percent().

See also

Examples

data <- get_acs_tables(
  geography = "county",
  county = "Baltimore city",
  state = "MD",
  table = "B08134"
)
#> ℹ Downloading table B08134
#> Warning: • You have not set a Census API key. Users without a key are limited to 500
#> queries per day and may experience performance limitations.
#> ℹ For best results, get a Census API key at
#> http://api.census.gov/data/key_signup.html and then supply the key to the
#> `census_api_key()` function to use it throughout your tidycensus session.
#> This warning is displayed once per session.
#> ✔ Downloading table B08134 [12.6s]
#> 
#> ℹ Labelling data
#> ✔ Caching table metadata
#> ℹ Labelling data
#> ✔ Caching column metadata
#> ℹ Labelling data
#> ✔ Labelling data [1.3s]
#> 

tbl_data <- filter_acs(data, indent == 1, line_number <= 10)
tbl_data <- select_acs_cols(tbl_data)

gt_acs(
  tbl_data,
  column_title_label = "Commute time",
  table = "B08134"
)
#> <div id="qpbqbvevvf" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#qpbqbvevvf table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #qpbqbvevvf thead, #qpbqbvevvf tbody, #qpbqbvevvf tfoot, #qpbqbvevvf tr, #qpbqbvevvf td, #qpbqbvevvf th {
#>   border-style: none;
#> }
#> 
#> #qpbqbvevvf p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #qpbqbvevvf .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #qpbqbvevvf .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #qpbqbvevvf .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #qpbqbvevvf .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #qpbqbvevvf .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #qpbqbvevvf .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #qpbqbvevvf .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #qpbqbvevvf .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #qpbqbvevvf .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #qpbqbvevvf .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #qpbqbvevvf .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #qpbqbvevvf .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #qpbqbvevvf .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #qpbqbvevvf .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #qpbqbvevvf .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qpbqbvevvf .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #qpbqbvevvf .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #qpbqbvevvf .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #qpbqbvevvf .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qpbqbvevvf .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #qpbqbvevvf .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qpbqbvevvf .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #qpbqbvevvf .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qpbqbvevvf .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qpbqbvevvf .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qpbqbvevvf .gt_left {
#>   text-align: left;
#> }
#> 
#> #qpbqbvevvf .gt_center {
#>   text-align: center;
#> }
#> 
#> #qpbqbvevvf .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #qpbqbvevvf .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #qpbqbvevvf .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #qpbqbvevvf .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #qpbqbvevvf .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #qpbqbvevvf .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #qpbqbvevvf .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #qpbqbvevvf .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #qpbqbvevvf .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #qpbqbvevvf .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #qpbqbvevvf .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #qpbqbvevvf .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="Commute time">Commute time</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="Est.">Est.</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="% share">% share</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="column_title" class="gt_row gt_left">Less than 10 minutes</td>
#> <td headers="estimate" class="gt_row gt_right">16,296 ± 895</td>
#> <td headers="perc_estimate" class="gt_row gt_right">7% ± 0%</td></tr>
#>     <tr><td headers="column_title" class="gt_row gt_left">10 to 14 minutes</td>
#> <td headers="estimate" class="gt_row gt_right">21,181 ± 1,146</td>
#> <td headers="perc_estimate" class="gt_row gt_right">9% ± 0%</td></tr>
#>     <tr><td headers="column_title" class="gt_row gt_left">15 to 19 minutes</td>
#> <td headers="estimate" class="gt_row gt_right">35,915 ± 1,531</td>
#> <td headers="perc_estimate" class="gt_row gt_right">15% ± 1%</td></tr>
#>     <tr><td headers="column_title" class="gt_row gt_left">20 to 24 minutes</td>
#> <td headers="estimate" class="gt_row gt_right">39,978 ± 1,823</td>
#> <td headers="perc_estimate" class="gt_row gt_right">16% ± 1%</td></tr>
#>     <tr><td headers="column_title" class="gt_row gt_left">25 to 29 minutes</td>
#> <td headers="estimate" class="gt_row gt_right">16,607 ± 1,037</td>
#> <td headers="perc_estimate" class="gt_row gt_right">7% ± 0%</td></tr>
#>     <tr><td headers="column_title" class="gt_row gt_left">30 to 34 minutes</td>
#> <td headers="estimate" class="gt_row gt_right">43,503 ± 1,748</td>
#> <td headers="perc_estimate" class="gt_row gt_right">18% ± 1%</td></tr>
#>     <tr><td headers="column_title" class="gt_row gt_left">35 to 44 minutes</td>
#> <td headers="estimate" class="gt_row gt_right">18,634 ± 1,018</td>
#> <td headers="perc_estimate" class="gt_row gt_right">8% ± 0%</td></tr>
#>     <tr><td headers="column_title" class="gt_row gt_left">45 to 59 minutes</td>
#> <td headers="estimate" class="gt_row gt_right">20,986 ± 1,314</td>
#> <td headers="perc_estimate" class="gt_row gt_right">9% ± 1%</td></tr>
#>     <tr><td headers="column_title" class="gt_row gt_left">60 or more minutes</td>
#> <td headers="estimate" class="gt_row gt_right">29,323 ± 1,492</td>
#> <td headers="perc_estimate" class="gt_row gt_right">12% ± 1%</td></tr>
#>   </tbody>
#>   <tfoot class="gt_sourcenotes">
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3">Source: 2017-2021 ACS 5-year Estimates, Table B08134.</td>
#>     </tr>
#>   </tfoot>
#>   
#> </table>
#> </div>