Create a gt table with values compared by name, geography, or variable

gt_acs_compare() is a variant of gt_acs() that uses pivot_acs_wider() to support comparisons of multiple named areas or multiple geographies side-by-side in a combined gt table.

Usage

gt_acs_compare(
  data,
  name_col = "NAME",
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  variable_col = "variable",
  column_title_col = "column_title",
  value_label = "Est.",
  moe_label = "MOE",
  perc_value_label = "% share",
  perc_moe_label = "% MOE",
  column_title_label = NULL,
  id_cols = column_title_col,
  id_expand = FALSE,
  names_from = name_col,
  values_from = NULL,
  names_vary = "slowest",
  names_glue = NULL,
  names_sep = "_",
  decimals = 0,
  currency_value = FALSE,
  merge_moe = TRUE,
  split = "last",
  limit = 1,
  reverse = TRUE,
  source_note = NULL,
  append_note = FALSE,
  hide_na_cols = TRUE,
  survey = "acs5",
  year = 2021,
  table = NULL,
  prefix = "Source: ",
  end = ".",
  use_md = FALSE,
  use_spanner = TRUE,
  ...
)

gt_acs_compare_vars(
  data,
  name_col = "NAME",
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  variable_col = "variable",
  column_title_col = "column_title",
  value_label = NULL,
  moe_label = "MOE",
  id_cols = name_col,
  names_from = variable_col,
  values_from = c(value_col, moe_col),
  use_spanner = FALSE,
  ...
)

Arguments

data

A data frame to pivot.

name_col

Name column. Defaults to "NAME". Ignored if names_from is not set to name_col.

value_col

Value column name. Defaults to "estimate"

moe_col

Margin of error column name. Defaults to "moe"

perc_prefix

Prefix string for percent value columns.

perc_sep

Separator string between perc_prefix and the value_col and moe_col strings.

variable_col

Variable column name. Defaults to "variable". Typically, dropped from table by default.

column_title_col, column_title_label

Column title column name and label. Defaults to "column_title" and NULL.

value_label

Value column label. Defaults to "Est.".

moe_label

Margin of error column label. Defaults to "MOE".

perc_value_label

Percent value column label.

perc_moe_label

Percent margin of error column label.

id_cols

Defaults to column_title_col. See tidyr::pivot_longer() for details.

id_expand

Should the values in the id_cols columns be expanded by expand() before pivoting? This results in more rows, the output will contain a complete expansion of all possible values in id_cols. Implicit factor levels that aren't represented in the data will become explicit. Additionally, the row values corresponding to the expanded id_cols will be sorted.

names_from, values_from

<tidy-select> A pair of arguments describing which column (or columns) to get the name of the output column (names_from), and which column (or columns) to get the cell values from (values_from).

If values_from contains multiple values, the value will be added to the front of the output column.

names_vary

When names_from identifies a column (or columns) with multiple unique values, and multiple values_from columns are provided, in what order should the resulting column names be combined?

"fastest" varies names_from values fastest, resulting in a column naming scheme of the form: value1_name1, value1_name2, value2_name1, value2_name2. This is the default.

"slowest" varies names_from values slowest, resulting in a column naming scheme of the form: value1_name1, value2_name1, value1_name2, value2_name2.

names_glue

Instead of names_sep and names_prefix, you can supply a glue specification that uses the names_from columns (and special .value) to create custom column names.

names_sep

If names_from or values_from contains multiple variables, this will be used to join their values together into a single string to use as a column name.

decimals

Number of decimal places

scalar<numeric|integer>(val>=0) // default: 2

This corresponds to the exact number of decimal places to use. A value such as 2.34 can, for example, be formatted with 0 decimal places and it would result in "2". With 4 decimal places, the formatted value becomes "2.3400". The trailing zeros can be removed with drop_trailing_zeros = TRUE. If you always need decimals = 0, the fmt_integer() function should be considered.

currency_value

If TRUE, use gt::fmt_currency() to format value columns instead of gt::fmt_number().

merge_moe

If TRUE, use gt::cols_merge_uncert() to merge the value_col and moe_col and the percent value and margin of error columns.

split

Splitting side

singl-kw:[last|first] // default: "last"

Should the delimiter splitting occur from the "last" instance of the delim character or from the "first"? The default here uses the "last" keyword, and splitting begins at the last instance of the delimiter in the column name. This option only has some consequence when there is a limit value applied that is lesser than the number of delimiter characters for a given column name (i.e., number of splits is not the maximum possible number).

limit

Limit for splitting

scalar<numeric|integer|character> // default: NULL (optional)

An optional limit to place on the splitting procedure. The default NULL means that a column name will be split as many times are there are delimiter characters. In other words, the default means there is no limit. If an integer value is given to limit then splitting will cease at the iteration given by limit. This works in tandem with split since we can adjust the number of splits from either the right side (split = "last") or left side (split = "first") of the column name.

reverse

Reverse vector of split names

scalar<logical> // default: FALSE

Should the order of split names be reversed? By default, this is FALSE.

source_note

Source note text

scalar<character> // required

Text to be used in the source note. We can optionally use the md() and html() functions to style the text as Markdown or to retain HTML elements in the text.

append_note

If TRUE, add source_note to the end of the generated ACS data label. If FALSE, any supplied source_note will be used instead of an ACS label.

hide_na_cols

If TRUE (default), hide any columns with all NA values using gt::cols_hide().

survey

ACS survey, "acs5", "acs3", or "acs1".

year

Based on the year and survey, acs_survey_ts() returns a vector of years for non-overlapping ACS samples to allow comparison.

table

One or more table IDs to include in label or source note.

prefix

Text to insert before ACS survey label.

end

A character string appended to the end of the full label. Defaults to ".".

use_md

If TRUE, pass source_note to gt::md() first.

...

Additional arguments passed on to methods.

See also