Format American Community Survey estimate and percent estimate columns for a gt table

Create or format a gt table with an estimate and margin of error and (optionally) percent estimate and margin of error value.

Usage

gt_census_cols(
  gt_object,
  est_cols = c("estimate", "moe"),
  est_col_label = "Est.",
  perc_cols = c("perc_estimate", "perc_moe"),
  perc_col_label = "% share",
  est_spanner = NULL,
  perc_spanner = NULL,
  combined_spanner = NULL,
  source_note = NULL,
  survey = "acs5",
  year = 2021,
  prefix = "Source: ",
  before = "",
  after = ".",
  tables = NULL,
  decimals = 0,
  column_title_col = "column_title",
  ...
)

Arguments

gt_object

If gt_object is a data frame, it is converted to a gt table with gt::gt().

est_cols, est_col_label, est_spanner

Passed to columns, col_labels, and spanner parameter for fmt_acs_estimate()

perc_cols, perc_col_label, perc_spanner

Passed to columns, col_labels, and spanner parameter for fmt_acs_percent()

combined_spanner

If not NULL, combined_spanner is passed to label parameter of gt::tab_spanner() using both est_cols and perc_cols as the columns parameter.

source_note

Source note text

scalar<character> // required

Text to be used in the source note. We can optionally use the md() and html() functions to style the text as Markdown or to retain HTML elements in the text.

survey

ACS survey, "acs5", "acs3", or "acs1".

year

Based on the year and survey, acs_survey_ts() returns a vector of years for non-overlapping ACS samples to allow comparison.

prefix

Text to insert before ACS survey label.

before, after

A character string to be added before/after each word.

tables

One or more table IDs to include in label or source note.

decimals

Number of decimal places

scalar<numeric|integer>(val>=0) // default: 2

This corresponds to the exact number of decimal places to use. A value such as 2.34 can, for example, be formatted with 0 decimal places and it would result in "2". With 4 decimal places, the formatted value becomes "2.3400". The trailing zeros can be removed with drop_trailing_zeros = TRUE. If you always need decimals = 0, the fmt_integer() function should be considered.

column_title_col

Column name to check for in the input gt table. This parameter may not be needed so may be removed.

...

Additional parameters passed to gt::fmt_number() by fmt_acs_estimate() or to gt::fmt_percent() by fmt_acs_percent().

See also