Join parent column titles to ACS data based on parent column ID values

join_acs_parent_column() uses data labelled with parent_column_id values to join parent column titles to a data frame of ACS data.

Usage

join_acs_parent_column(
  data,
  column_id_col = "column_id",
  column_title_col = "column_title",
  parent_id_col = "parent_column_id",
  suffix = c("", "_parent"),
  na_matches = "never",
  relationship = "many-to-one"
)

Arguments

data

A data frame with the specified column names. Expected to be labelled using label_acs_metadata().

column_id_col, column_title_col, parent_id_col

Column ID, column title, and parent column ID.

suffix

Suffix passed to dplyr::left_join(), Default: c("", "_parent")

na_matches

Should two NA or two NaN values match?

"na", the default, treats two NA or two NaN values as equal, like %in%, match(), and merge().

"never" treats two NA or two NaN values as different, and will never match them together or to any other values. This is similar to joins for database sources and to base::merge(incomparables = NA).

relationship

Handling of the expected relationship between the keys of x and y. If the expectations chosen from the list below are invalidated, an error is thrown.

NULL, the default, doesn't expect there to be any relationship between x and y. However, for equality joins it will check for a many-to-many relationship (which is typically unexpected) and will warn if one occurs, encouraging you to either take a closer look at your inputs or make this relationship explicit by specifying "many-to-many".

See the Many-to-many relationships section for more details.

"one-to-one" expects:

Each row in x matches at most 1 row in y.

Each row in y matches at most 1 row in x.

"one-to-many" expects:

Each row in y matches at most 1 row in x.

"many-to-one" expects:

Each row in x matches at most 1 row in y.

"many-to-many" doesn't perform any relationship checks, but is provided to allow you to be explicit about this relationship if you know it exists.

relationship doesn't handle cases where there are zero matches. For that, see unmatched.

Value

A data frame with added parent column title.