Add a Census data source note to a gt table

tab_acs_source_note() adds a source note to a gt table using acs_survey_label_table() and gt::tab_source_note().

Usage

tab_acs_source_note(
  gt_object,
  source_note = NULL,
  append_note = FALSE,
  survey = "acs5",
  year = 2021,
  table = NULL,
  table_label = "Table",
  prefix = "Source: ",
  end = ".",
  use_md = FALSE,
  ...
)

Arguments

gt_object

A gt object.

source_note

Source note text

scalar<character> // required

Text to be used in the source note. We can optionally use the md() and html() functions to style the text as Markdown or to retain HTML elements in the text.

append_note

If TRUE, add source_note to the end of the generated ACS data label. If FALSE, any supplied source_note will be used instead of an ACS label.

survey

ACS survey, "acs5", "acs3", or "acs1".

year

Based on the year and survey, acs_survey_ts() returns a vector of years for non-overlapping ACS samples to allow comparison.

table

One or more table IDs to include in label or source note.

table_label

Label to use when referring to table or tables. A "s" is appended to the end of the table_label if tables is more than length 1.

prefix

Text to insert before ACS survey label.

end

A character string appended to the end of the full label. Defaults to ".".

use_md

If TRUE, pass source_note to gt::md() first.

...

For tab_acs_source_note(), additional parameters passed to acs_survey_label_table(). For cols_merge_uncert_ext(), additional parameters passed to gt::cols_merge_uncert(). For fmt_acs_percent(), additional parameters passed to gt::fmt_percent().

See also