Create new spanners using gt::tab_spanner_delim()

Create new spanners using gt::tab_spanner_delim()

Usage

tab_acs_spanner_delim(
  data,
  column_title_col = "column_title",
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  columns = starts_with(c(value_col, moe_col, perc_prefix)),
  delim = "_",
  split = "last",
  limit = 1,
  reverse = TRUE,
  ...
)

Arguments

data

The gt table data object

obj:<gt_tbl> // required

This is the gt table object that is commonly created through use of the gt() function.

columns

Columns to target

<column-targeting expression> // default: everything()

The columns to consider for the splitting, relabeling, and spanner setting operations. Can either be a series of column names provided in c(), a vector of column indices, or a select helper function. Examples of select helper functions include starts_with(), ends_with(), contains(), matches(), one_of(), num_range(), and everything(). This argument works in tandem with the spanners argument.

delim

Delimiter for splitting

scalar<character> // required

The delimiter to use to split an input column name. This should be a single character (e.g., "_", ".", etc.).

split

Splitting side

singl-kw:[last|first] // default: "last"

Should the delimiter splitting occur from the "last" instance of the delim character or from the "first"? The default here uses the "last" keyword, and splitting begins at the last instance of the delimiter in the column name. This option only has some consequence when there is a limit value applied that is lesser than the number of delimiter characters for a given column name (i.e., number of splits is not the maximum possible number).

limit

Limit for splitting

scalar<numeric|integer|character> // default: NULL (optional)

An optional limit to place on the splitting procedure. The default NULL means that a column name will be split as many times are there are delimiter characters. In other words, the default means there is no limit. If an integer value is given to limit then splitting will cease at the iteration given by limit. This works in tandem with split since we can adjust the number of splits from either the right side (split = "last") or left side (split = "first") of the column name.

reverse

Reverse vector of split names

scalar<logical> // default: FALSE

Should the order of split names be reversed? By default, this is FALSE.