Create a gt table with formatted ACS estimate and percent estimate columns

Create or format a gt table with an estimate and margin of error and (optionally) percent estimate and margin of error value. Use in combination with the select_acs() helper function to prep data before creating a table.

Usage

gt_acs(
  data,
  rownames_to_stub = FALSE,
  row_group_as_column = FALSE,
  ...,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  column_title_col = "column_title",
  name_col = "NAME",
  perc_value_label = "% share",
  value_label = "Est.",
  column_title_label = NULL,
  name_label = NULL,
  est_spanner = NULL,
  perc_spanner = NULL,
  combined_spanner = NULL,
  decimals = 0,
  source_note = NULL,
  append_note = FALSE,
  drop_geometry = TRUE,
  hide_na_cols = TRUE,
  currency_value = FALSE,
  survey = "acs5",
  year = 2021,
  table = NULL,
  prefix = "Source: ",
  end = ".",
  est_cols = NULL,
  perc_cols = NULL
)

Arguments

data

Input data table

obj:<data.frame>|obj:<tbl_df> // required

A data.frame object or a tibble (tbl_df).

rownames_to_stub

Use data frame row labels in the stub

scalar<logical> // default: FALSE

An option to take rownames from the input data table (should they be available) as row labels in the display table stub.

row_group_as_column

Mode for displaying row group labels in the stub

scalar<logical> // default: FALSE

An option that alters the display of row group labels. By default this is FALSE and row group labels will appear in dedicated rows above their respective groups of rows. If TRUE row group labels will occupy a secondary column in the table stub.

...

Additional parameters passed to gt::fmt_number() by fmt_acs_estimate() or to gt::fmt_percent() by fmt_acs_percent().

value_col

Value column name. Defaults to "estimate"

moe_col

Margin of error column name. Defaults to "moe"

perc_prefix

Prefix string for percent value columns.

perc_sep

Separator string between perc_prefix and the value_col and moe_col strings.

column_title_col, column_title_label

Column title and label. If column_title_label is a string, column_title_col is required. column_title_label can also be a named vector in the format of c("label" = "column"). column_title_col defaults to "column_title". If column_title_label is "from_table", the label is set based on the simple_table_title column in the table metadata.

name_col, name_label

Place name column and label. name_label can be a string or a named vector (similar to column_title_label). name_col defaults to "NAME"

perc_value_label

Percent value column label.

value_label

Value column label. Defaults to "Est.".

est_spanner, perc_spanner

Spanner labels for estimate and percent estimate columns.

combined_spanner

If not NULL, combined_spanner is passed to label parameter of gt::tab_spanner() using the value columns and percent columns as the columns parameter.

decimals

Number of decimal places

scalar<numeric|integer>(val>=0) // default: 2

This corresponds to the exact number of decimal places to use. A value such as 2.34 can, for example, be formatted with 0 decimal places and it would result in "2". With 4 decimal places, the formatted value becomes "2.3400". The trailing zeros can be removed with drop_trailing_zeros = TRUE. If you always need decimals = 0, the fmt_integer() function should be considered.

source_note

Source note text

scalar<character> // required

Text to be used in the source note. We can optionally use the md() and html() functions to style the text as Markdown or to retain HTML elements in the text.

append_note

If TRUE, add source_note to the end of the generated ACS data label. If FALSE, any supplied source_note will be used instead of an ACS label.

drop_geometry

If TRUE (default) and data is an sf object, drop geometry before turning the data frame into a table.

hide_na_cols

If TRUE (default), hide columns where all values are NA.

currency_value

If TRUE, use gt::fmt_currency() to format value columns instead of gt::fmt_number().

survey

ACS survey, "acs5", "acs3", or "acs1".

year

Based on the year and survey, acs_survey_ts() returns a vector of years for non-overlapping ACS samples to allow comparison.

table

One or more table IDs to include in label or source note.

prefix

Text to insert before ACS survey label.

end

A character string appended to the end of the full label. Defaults to ".".

est_cols, perc_cols

Deprecated. Estimate and percent estimate columns.

See also

Examples

if (FALSE) {
if (interactive()) {
  data <- get_acs_tables(
    geography = "county",
    county = "Baltimore city",
    state = "MD",
    table = "B08134"
  )

  tbl_data <- filter_acs(data, indent == 1, line_number <= 10)
  tbl_data <- select_acs(tbl_data)

  gt_acs(
    tbl_data,
    column_title_label = "Commute time",
    table = "B08134"
  )
}
}